requirements.txt
tensorflow-gpu==1.13.1 
numpy 
matplotlib 
pillow 
opencv-python 
keras
scikit-learn
tqdm